# lib/api-zod/src/generated/types/createTripBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TownRide Carpooling API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTripBody {
  origin: string;
  destination: string;
  date: string;
  time: string;
  availableSeats: number;
  pricePerSeat: number;
  meetingPoint: string;
  acceptsPackages?: boolean;
}